feat: #22 #23 update UI 플레이어, 앨범창 UI 업데이트
