+++ Music-Player/Music-Player/View/ContentView.swift
@@ -8,33 +8,6 @@
 import SwiftUI
 import MediaPlayer
 
-class AuthViewModel: ObservableObject {
-    @Published var authStatus: Bool = true
-    
-    init() {
-        getAuthrization()
-    }
-    
-    func getAuthrization()  {
-        DispatchQueue.main.async {
-            let status = MPMediaLibrary.authorizationStatus()
-            if(status == MPMediaLibraryAuthorizationStatus.authorized){
-                self.authStatus = true
-            }else{
-                MPMediaLibrary.requestAuthorization() { status in
-                    DispatchQueue.main.async {
-                        if status == .authorized {
-                            self.authStatus = true
-                        }else{
-                            self.authStatus = false
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
 struct ContentView: View {
     
     @State private var player = MPMusicPlayerController.applicationMusicPlayer
